consistency_check/requirements.txt
connexion[flask,uvicorn]
flask_cors
requests
pyyaml
apscheduler
